Extract useLanguageFilter hook to DRY up timeline routes
The five timeline routes (fediverse, local, feed, feed/articles,
feed/without-shares) each duplicated the same three-line block to read
the ?language= search param and derive activeLanguage / initialLang,
plus an identical six-line buildHref inline that differed only in the
base path string.

Extract a useLanguageFilter(basePath) hook in src/lib/useLanguageFilter.ts
that encapsulates useSearchParams, useLocation, and the buildHref logic.
Each route now calls the hook with its path and spreads the three
returned values, removing all duplicated code.

https://github.com/hackers-pub/hackerspub/pull/302#discussion_r3266713676
